refactor(app): getAttachedProtocolModuleMatches to xforms
