+++ app/src/organisms/Devices/ProtocolRun/SetupLiquids/__tests__/SetupLiquidsMap.test.tsx
@@ -24,7 +24,7 @@ import { i18n } from '/app/i18n'
 import { useAttachedModules } from '/app/resources/modules'
 import { LabwareInfoOverlay } from '../../LabwareInfoOverlay'
 import { getStandardDeckViewLayerBlockList } from '/app/local-resources/deck_configuration'
-import { getAttachedProtocolModuleMatches } from '/app/organisms/ODD/ProtocolSetup/ProtocolSetupModulesAndDeck/utils'
+import { getAttachedProtocolModuleMatches } from '/app/transformations/analysis'
 import {
   getProtocolModulesInfo,
   getLabwareRenderInfo,
@@ -52,7 +52,7 @@ vi.mock('@opentrons/components', async importOriginal => {
 vi.mock('@opentrons/components/src/hardware-sim/BaseDeck')
 vi.mock('../../LabwareInfoOverlay')
 vi.mock('/app/resources/modules')
-vi.mock('../../../../ODD/ProtocolSetup/ProtocolSetupModulesAndDeck/utils')
+vi.mock('/app/transformations/analysis')
 vi.mock('/app/transformations/analysis')
 vi.mock('/app/resources/deck_configuration/utils')
 vi.mock('@opentrons/shared-data', async importOriginal => {
