+++ app/src/transformations/analysis/index.ts
@@ -3,3 +3,4 @@ export * from './getProtocolModulesInfo'
 export * from './getLabwareOffsetLocation'
 export * from './getLabwareRenderInfo'
 export * from './liquids'
+export * from './getAttachedProtocolModuleMatches'
